Broke candidate and employer login forms into two child components of Login component. Renamed document title to Hire Me.
