+++ package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "apollo-server-express": "^2.15.0",
-    "graphql": "^15.8.0"
+    "graphql": "^15.8.0",
+    "mongodb": "^6.1.0"
   }
 }
